requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.2
django-admin-honeypot @ git+https://github.com/dmpayton/django-admin-honeypot.git@a840496d183df89965eeb33c9a5dd9ea3919dfff
django-cors-headers==3.13.0
django-filter==22.1
django-ipware==4.0.2
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
dulwich==0.20.46
Faker==4.14.2
filelock==3.8.0
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
Pillow==9.3.0
pipenv==2022.10.12
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.1.2
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2020.4
redis==4.3.4
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tomlkit==0.11.5
urllib3==1.26.12
validate-docbr==1.8.2
virtualenv==20.16.6
virtualenv-clone==0.5.7
webencodings==0.5.1
wrapt==1.14.1
